Update getting_started_hassio.rst
Hi, i added some information for users how use HA as a VM e.g. Docker. In this setup the ESPHome dashboard is not available within HA. It is only possible to integrate nodes that are already preconfigured for ESPHome. If this information is correct. thx for your create work.